Lab6 arreglado

+++ app/src/main/java/dam/isi/frsf/utn/edu/ar/lab05/dao/TareaDAO.java
@@ -9,6 +9,9 @@ import android.util.Log;
 import java.util.ArrayList;
 import java.util.List;
 
+import dam.isi.frsf.utn.edu.ar.lab05.Exception.ProyectoException;
+import dam.isi.frsf.utn.edu.ar.lab05.Exception.TareaException;
+import dam.isi.frsf.utn.edu.ar.lab05.MyApplication;
 import dam.isi.frsf.utn.edu.ar.lab05.modelo.Prioridad;
 import dam.isi.frsf.utn.edu.ar.lab05.modelo.Proyecto;
 import dam.isi.frsf.utn.edu.ar.lab05.modelo.Tarea;
@@ -36,29 +39,29 @@ public class TareaDAO {
             ProyectoDBMetadata.TABLA_TAREAS_ALIAS+"."+ProyectoDBMetadata.TablaTareasMetadata.PRIORIDAD+" = "+ProyectoDBMetadata.TABLA_PRIORIDAD_ALIAS+"."+ProyectoDBMetadata.TablaPrioridadMetadata._ID +" AND "+
             ProyectoDBMetadata.TABLA_TAREAS_ALIAS+"."+ProyectoDBMetadata.TablaTareasMetadata.PROYECTO+" = ?";
 
-
+    private static TareaDAO ourInstance = new TareaDAO();
     private static final int MODO_PERSISTENCIA_MIXTA = 2;  // Los datos se almacenan en la api rest y en local
     private static final int MODO_PERSISTENCIA_LOCAL = 1;  // Los datos se almacenan solamente en la bdd local
     private static final int MODO_PERSISTENCIA_REMOTA = 0; // Los datos se almacenan solamente en la nube
     private static int MODO_PERSISTENCIA_CONFIGURADA; // Como default es remota
-    private ProyectoOpenHelper dbHelper;
+    private static boolean usarApiRest = true; // default true
+    private static ProyectoOpenHelper dbHelper = new ProyectoOpenHelper(MyApplication.getAppContext());
+    private static ProyectoApiRest daoApiRest = new ProyectoApiRest();
+    private static final UsuarioDAO daoUsuario = UsuarioDAO.getInstance();
+    private static final ProyectoDAO daoProyecto = ProyectoDAO.getInstance();
     private SQLiteDatabase db;
     private List<Usuario> listaUsuarios;
-    private ProyectoApiRest daoApiRest;
-    private UsuarioDAO daoUsuario;
-    private ProyectoDAO daoProyecto;
-    private static boolean usarApiRest;
-    private Context context;
 
 
-    public TareaDAO(Context c) {
-        MODO_PERSISTENCIA_CONFIGURADA = MODO_PERSISTENCIA_REMOTA;
-        this.context=c;
-        this.dbHelper = new ProyectoOpenHelper(c);
-        this.daoApiRest = new ProyectoApiRest();
-        this.daoProyecto = new ProyectoDAO(c);
-        this.daoUsuario = new UsuarioDAO(c);
+    private TareaDAO(){
+
     }
+
+    public static TareaDAO getInstance(){
+        MODO_PERSISTENCIA_CONFIGURADA = MODO_PERSISTENCIA_REMOTA;
+        return ourInstance;
+    }
+
 
     public void open(){
         this.open(false);
@@ -77,38 +80,38 @@ public class TareaDAO {
         db = dbHelper.getReadableDatabase();
     }
 
-    public Tarea getTarea(int idTarea)
-    {
+    public Tarea getTarea(int idTarea) throws TareaException {
         Tarea nuevaTarea = null;
-        Cursor resultadoTareas;
-        try
-        {
-            open(false);
-            resultadoTareas = db.rawQuery("SELECT * " + " FROM "+ProyectoDBMetadata.TABLA_TAREAS+" WHERE "+ProyectoDBMetadata.TablaTareasMetadata._ID+" = "+idTarea,null);
-            resultadoTareas.moveToFirst();
-
-            String descripcion = resultadoTareas.getString(1);
-            Integer horasEstimadas = resultadoTareas.getInt(2);
-            Integer minutosTrabajados = resultadoTareas.getInt(3);
-            Prioridad prioridad = getPrioridad(resultadoTareas.getInt(4));
-            Usuario responsable = daoUsuario.getUsuario(resultadoTareas.getInt(5));
-            Proyecto proyecto = daoProyecto.getProyecto(resultadoTareas.getInt(6));
-            Boolean finalizada;
-            if(resultadoTareas.getInt(7) == 0 )
-            {
-                finalizada=false;
+        try {
+            if (usarApiRest) {
+                 nuevaTarea = daoApiRest.getTarea(idTarea);
             }
             else {
-                finalizada = true;
+                Cursor resultadoTareas;
+
+                open(false);
+                resultadoTareas = db.rawQuery("SELECT * " + " FROM " + ProyectoDBMetadata.TABLA_TAREAS + " WHERE " + ProyectoDBMetadata.TablaTareasMetadata._ID + " = " + idTarea, null);
+                resultadoTareas.moveToFirst();
+
+                String descripcion = resultadoTareas.getString(1);
+                Integer horasEstimadas = resultadoTareas.getInt(2);
+                Integer minutosTrabajados = resultadoTareas.getInt(3);
+                Prioridad prioridad = getPrioridad(resultadoTareas.getInt(4));
+                Usuario responsable = daoUsuario.getUsuario(resultadoTareas.getInt(5));
+                Proyecto proyecto = daoProyecto.getProyecto(resultadoTareas.getInt(6));
+                Boolean finalizada;
+                if (resultadoTareas.getInt(7) == 0) {
+                    finalizada = false;
+                } else {
+                    finalizada = true;
+                }
+                nuevaTarea = new Tarea(idTarea, finalizada, horasEstimadas, minutosTrabajados, proyecto, prioridad, responsable, descripcion);
+                resultadoTareas.close();
             }
-            nuevaTarea = new Tarea(idTarea,finalizada,horasEstimadas,minutosTrabajados,proyecto,prioridad,responsable,descripcion);
-            resultadoTareas.close();
         }
         catch(Exception e)
         {
-
-            System.out.println("Exploto la bd al buscar una tarea");
-
+            throw new TareaException("La tarea no pudo ser encontrada");
         }
         return nuevaTarea;
     }
@@ -117,28 +120,39 @@ public class TareaDAO {
      * Borra tarea por id
      * @param idTarea
      */
-    public void borrarTarea(int idTarea)
-    {
-        String[] args = { String.valueOf(idTarea) };
-        try
-        {
-            open(true);
-            db.delete(ProyectoDBMetadata.TABLA_TAREAS,"_id=?", args);
+    public void borrarTarea(int idTarea) throws TareaException {
+        try {
+            if (usarApiRest) {
+                daoApiRest.borrarTarea(idTarea);
+            }
+            else {
+                String[] args = {String.valueOf(idTarea)};
+                open(true);
+                db.delete(ProyectoDBMetadata.TABLA_TAREAS, "_id=?", args);
+            }
         }
-        catch(Exception e)
-        {
-            System.out.println(e.getMessage());
-            System.out.println("Bd exploto al eliminar tarea");
+        catch(Exception e){
+            throw new TareaException("La tarea no pudo ser eliminada");
         }
     }
 
-    public void finalizar(Integer idTarea){
-        //Establecemos los campos-valores a actualizar
-        ContentValues valores = new ContentValues();
-        valores.put(ProyectoDBMetadata.TablaTareasMetadata.FINALIZADA,1);
-        SQLiteDatabase mydb =dbHelper.getWritableDatabase();
-        mydb.update(ProyectoDBMetadata.TABLA_TAREAS, valores, "_id=?", new String[]{idTarea.toString()});
-        // mydb.close();
+    public void finalizar(Integer idTarea) throws TareaException{
+        try {
+            if (usarApiRest) {
+                daoApiRest.finalizarTarea(idTarea);
+            }
+            else {
+                //Establecemos los campos-valores a actualizar
+                ContentValues valores = new ContentValues();
+                valores.put(ProyectoDBMetadata.TablaTareasMetadata.FINALIZADA, 1);
+                SQLiteDatabase mydb = dbHelper.getWritableDatabase();
+                mydb.update(ProyectoDBMetadata.TABLA_TAREAS, valores, "_id=?", new String[]{idTarea.toString()});
+                // mydb.close();
+            }
+        }
+        catch(Exception e){
+            throw new TareaException("La tarea no pudo marcarse como finalizada");
+        }
     }
 
     public List<Tarea> listarDesviosPlanificacion(Boolean soloTerminadas,Integer desvioMinimo){
@@ -230,58 +244,82 @@ public class TareaDAO {
         }
     }
 
-    public void actualizarTarea(Tarea t){
-        ContentValues datosAGuardar = new ContentValues();
-        datosAGuardar.put(ProyectoDBMetadata.TablaTareasMetadata.HORAS_PLANIFICADAS,t.getHorasEstimadas());
-        datosAGuardar.put(ProyectoDBMetadata.TablaTareasMetadata.MINUTOS_TRABAJADOS,0);
-        datosAGuardar.put(ProyectoDBMetadata.TablaTareasMetadata.TAREA,t.getDescripcion());
-        datosAGuardar.put(ProyectoDBMetadata.TablaTareasMetadata.PRIORIDAD,t.getPrioridad().getId());
-        datosAGuardar.put(ProyectoDBMetadata.TablaTareasMetadata.RESPONSABLE,t.getResponsable().getId());
-        datosAGuardar.put(ProyectoDBMetadata.TablaTareasMetadata. PROYECTO,t.getProyecto().getId());
-        open(true);
+    public void actualizarTarea(Tarea t) throws TareaException {
         try {
-            db.update(ProyectoDBMetadata.TABLA_TAREAS,datosAGuardar,ProyectoDBMetadata.TablaTareasMetadata._ID+"="+t.getId(),null);
+            if (usarApiRest) {
+                daoApiRest.actualizarTarea(t);
+            }
+            else {
+                ContentValues datosAGuardar = new ContentValues();
+                datosAGuardar.put(ProyectoDBMetadata.TablaTareasMetadata.HORAS_PLANIFICADAS, t.getHorasEstimadas());
+                datosAGuardar.put(ProyectoDBMetadata.TablaTareasMetadata.MINUTOS_TRABAJADOS, 0);
+                datosAGuardar.put(ProyectoDBMetadata.TablaTareasMetadata.TAREA, t.getDescripcion());
+                datosAGuardar.put(ProyectoDBMetadata.TablaTareasMetadata.PRIORIDAD, t.getPrioridad().getId());
+                datosAGuardar.put(ProyectoDBMetadata.TablaTareasMetadata.RESPONSABLE, t.getResponsable().getId());
+                datosAGuardar.put(ProyectoDBMetadata.TablaTareasMetadata.PROYECTO, t.getProyecto().getId());
+                open(true);
+                db.update(ProyectoDBMetadata.TABLA_TAREAS, datosAGuardar, ProyectoDBMetadata.TablaTareasMetadata._ID + "=" + t.getId(), null);
+            }
         }
-        catch(Exception e)
-        {
-            System.out.println(e.getMessage());
-            System.out.println("BD Exploto al actualizar tarea");
+        catch(Exception e){
+            throw new TareaException("La tarea no pudo ser actualizada, intente nuevamente");
         }
     }
 
-    public void nuevaTarea(Tarea t)
-    {
-        ContentValues datosAGuardar = new ContentValues();
-        datosAGuardar.put(ProyectoDBMetadata.TablaTareasMetadata.HORAS_PLANIFICADAS,t.getHorasEstimadas());
-        datosAGuardar.put(ProyectoDBMetadata.TablaTareasMetadata.MINUTOS_TRABAJADOS,0);
-        datosAGuardar.put(ProyectoDBMetadata.TablaTareasMetadata.TAREA,t.getDescripcion());
-        datosAGuardar.put(ProyectoDBMetadata.TablaTareasMetadata.PRIORIDAD,t.getPrioridad().getId());
-        datosAGuardar.put(ProyectoDBMetadata.TablaTareasMetadata.RESPONSABLE,t.getResponsable().getId());
-        datosAGuardar.put(ProyectoDBMetadata.TablaTareasMetadata. PROYECTO,t.getProyecto().getId());
-        open(true);
+    public void nuevaTarea(Tarea t) throws TareaException {
         try {
-            db.insert(ProyectoDBMetadata.TABLA_TAREAS,null,datosAGuardar);
+            if (usarApiRest) {
+                daoApiRest.guardarTarea(t);
+            }
+            else {
+                ContentValues datosAGuardar = new ContentValues();
+                datosAGuardar.put(ProyectoDBMetadata.TablaTareasMetadata.HORAS_PLANIFICADAS, t.getHorasEstimadas());
+                datosAGuardar.put(ProyectoDBMetadata.TablaTareasMetadata.MINUTOS_TRABAJADOS, 0);
+                datosAGuardar.put(ProyectoDBMetadata.TablaTareasMetadata.TAREA, t.getDescripcion());
+                datosAGuardar.put(ProyectoDBMetadata.TablaTareasMetadata.PRIORIDAD, t.getPrioridad().getId());
+                datosAGuardar.put(ProyectoDBMetadata.TablaTareasMetadata.RESPONSABLE, t.getResponsable().getId());
+                datosAGuardar.put(ProyectoDBMetadata.TablaTareasMetadata.PROYECTO, t.getProyecto().getId());
+                open(true);
+                db.insert(ProyectoDBMetadata.TABLA_TAREAS, null, datosAGuardar);
+            }
         }
-        catch(Exception e)
-        {
-            System.out.println(e.getMessage());
-            System.out.println("BD Exploto en el insert de tarea");
+        catch(Exception e){
+            throw new TareaException("La tarea no pudo ser creada");
         }
     }
 
-    public Cursor listaTareas(Integer idProyecto){
-        Cursor cursorPry = db.rawQuery("SELECT "+ProyectoDBMetadata.TablaProyectoMetadata._ID+ " FROM "+ProyectoDBMetadata.TABLA_PROYECTO,null);
-      //  Integer idPry= 0;
-        Integer idPry=idProyecto;
-        /*
-        if(cursorPry.moveToFirst()){
-            idPry=cursorPry.getInt(0);
+    /**
+     * Devuelve un cursor con las tareas del proyecto parametro
+     * @param idProyecto
+     * @return
+     */
+    public Cursor listaTareas(Integer idProyecto) {
+        Cursor cursorTareas=null;
+        try{
+            if(usarApiRest){
+                cursorTareas=daoApiRest.getCursorTareas(idProyecto);
+                cursorTareas.moveToFirst();
+            }
+            else {
+                cursorTareas = db.rawQuery("SELECT " + ProyectoDBMetadata.TablaProyectoMetadata._ID + " FROM " + ProyectoDBMetadata.TABLA_PROYECTO, null);
+                //  Integer idPry= 0;
+                Integer idPry = idProyecto;
+                /*
+                if(cursorPry.moveToFirst()){
+                    idPry=cursorPry.getInt(0);
+                }
+                */
+                Cursor cursor = null;
+                Log.d("Listar Tareas", "PROYECTO ID: " + idPry.toString() + " - " + _SQL_TAREAS_X_PROYECTO);
+                cursor = db.rawQuery(_SQL_TAREAS_X_PROYECTO, new String[]{idPry.toString()});
+                return cursor;
+            }
         }
-        */
-        Cursor cursor = null;
-        Log.d("Listar Tareas","PROYECTO ID: "+idPry.toString()+" - "+ _SQL_TAREAS_X_PROYECTO);
-        cursor = db.rawQuery(_SQL_TAREAS_X_PROYECTO,new String[]{idPry.toString()});
-        return cursor;
+        catch(Exception e){
+            e.printStackTrace();
+           // throw new TareaException("No se encontraron tareas asociadas al proyecto");
+        }
+        return cursorTareas;
     }
 
     /**
@@ -301,11 +339,11 @@ public class TareaDAO {
 
         }
         catch (Exception e) {
-            System.out.println(e.getMessage());
-            System.out.println("Error al buscar una prioridad");
+            throw new TareaException("Se produjo une rror al encontrar la prioridad");
         }
         finally {
             return prioridad;
         }
     }
+
 }
